usb: refactor
