+++ src/machine/machine_nrf52840_usb_reset_none.go
@@ -3,5 +3,5 @@
 
 package machine
 
-func checkShouldReset() {
+func ResetProcessor() {
 }
