+++ src/machine/machine_nrf52840_usb_reset_uf2.go
@@ -14,12 +14,10 @@ const (
 	DFU_MAGIC_OTA_RESET         = 0xA8
 )
 
-// checkShouldReset is called by the USB-CDC implementation to check whether to
-// reset into the bootloader/OTA and if so, resets the chip appropriately.
-func checkShouldReset() {
-	if usbLineInfo.dwDTERate == 1200 && usbLineInfo.lineState&usb_CDC_LINESTATE_DTR == 0 {
-		EnterUF2Bootloader()
-	}
+// ResetProcessor should perform a system reset in preparation
+// to switch to the bootloader to flash new firmware.
+func ResetProcessor() {
+	EnterUF2Bootloader()
 }
 
 // EnterSerialBootloader resets the chip into the serial bootloader. After
