+++ src/machine/usb.go
@@ -8,24 +8,30 @@ import (
 	"runtime/volatile"
 )
 
-var usbDescriptor = descriptorCDC
+type USBDevice struct {
+}
 
 var (
-	errUSBCDCBufferEmpty      = errors.New("USB-CDC buffer empty")
-	errUSBCDCWriteByteTimeout = errors.New("USB-CDC write byte timeout")
-	errUSBCDCReadTimeout      = errors.New("USB-CDC read timeout")
-	errUSBCDCBytesRead        = errors.New("USB-CDC invalid number of bytes read")
+	USB    = &USBDevice{}
+	USBCDC Serialer
 )
 
-const cdcLineInfoSize = 7
-
-type cdcLineInfo struct {
-	dwDTERate   uint32
-	bCharFormat uint8
-	bParityType uint8
-	bDataBits   uint8
-	lineState   uint8
+type Serialer interface {
+	WriteByte(c byte) error
+	Write(data []byte) (n int, err error)
+	Configure(config UARTConfig) error
+	Buffered() int
+	ReadByte() (byte, error)
 }
+
+var usbDescriptor = descriptorCDC
+
+const (
+	usbDescriptorConfigCDC = 1 << iota
+	usbDescriptorConfigHID
+)
+
+var usbDescriptorConfig uint8 = usbDescriptorConfigCDC
 
 // strToUTF16LEDescriptor converts a utf8 string into a string descriptor
 // note: the following code only converts ascii characters to UTF16LE. In order
@@ -47,11 +53,35 @@ var (
 	usb_STRING_LANGUAGE = [2]uint16{(3 << 8) | (2 + 2), 0x0409} // English
 )
 
+const cdcLineInfoSize = 7
+
+var (
+	errUSBCDCBufferEmpty      = errors.New("USB-CDC buffer empty")
+	errUSBCDCWriteByteTimeout = errors.New("USB-CDC write byte timeout")
+	errUSBCDCReadTimeout      = errors.New("USB-CDC read timeout")
+	errUSBCDCBytesRead        = errors.New("USB-CDC invalid number of bytes read")
+)
+
+var (
+	usbEndpointDescriptors [8]usbDeviceDescriptor
+
+	udd_ep_control_cache_buffer [256]uint8
+	udd_ep_in_cache_buffer      [7][64]uint8
+	udd_ep_out_cache_buffer     [7][64]uint8
+
+	isEndpointHalt        = false
+	isRemoteWakeUpEnabled = false
+
+	usbConfiguration uint8
+	usbSetInterface  uint8
+)
+
 const (
 	usb_IMANUFACTURER = 1
 	usb_IPRODUCT      = 2
 	usb_ISERIAL       = 3
 
+	usb_ENDPOINT_TYPE_DISABLE     = 0xFF
 	usb_ENDPOINT_TYPE_CONTROL     = 0x00
 	usb_ENDPOINT_TYPE_ISOCHRONOUS = 0x01
 	usb_ENDPOINT_TYPE_BULK        = 0x02
@@ -102,8 +132,10 @@ const (
 	usb_CDC_ACM_INTERFACE  = 0 // CDC ACM
 	usb_CDC_DATA_INTERFACE = 1 // CDC Data
 	usb_CDC_FIRST_ENDPOINT = 1
+	usb_HID_INTERFACE      = 2 // HID
 
 	// Endpoint
+	usb_CONTROL_ENDPOINT = 0
 	usb_CDC_ENDPOINT_ACM = 1
 	usb_CDC_ENDPOINT_OUT = 2
 	usb_CDC_ENDPOINT_IN  = 3
@@ -128,37 +160,23 @@ const (
 	usb_REQUEST_DEVICETOHOST_CLASS_INTERFACE    = (usb_REQUEST_DEVICETOHOST | usb_REQUEST_CLASS | usb_REQUEST_INTERFACE)
 	usb_REQUEST_HOSTTODEVICE_CLASS_INTERFACE    = (usb_REQUEST_HOSTTODEVICE | usb_REQUEST_CLASS | usb_REQUEST_INTERFACE)
 	usb_REQUEST_DEVICETOHOST_STANDARD_INTERFACE = (usb_REQUEST_DEVICETOHOST | usb_REQUEST_STANDARD | usb_REQUEST_INTERFACE)
+)
 
-	// CDC Class requests
-	usb_CDC_SET_LINE_CODING        = 0x20
-	usb_CDC_GET_LINE_CODING        = 0x21
-	usb_CDC_SET_CONTROL_LINE_STATE = 0x22
-	usb_CDC_SEND_BREAK             = 0x23
+var (
+	callbackUSBTx    [8]func()
+	callbackUSBRx    [8]func([]byte)
+	callbackUSBSetup [3]func(USBSetup) bool
 
-	usb_CDC_V1_10                         = 0x0110
-	usb_CDC_COMMUNICATION_INTERFACE_CLASS = 0x02
-
-	usb_CDC_CALL_MANAGEMENT             = 0x01
-	usb_CDC_ABSTRACT_CONTROL_MODEL      = 0x02
-	usb_CDC_HEADER                      = 0x00
-	usb_CDC_ABSTRACT_CONTROL_MANAGEMENT = 0x02
-	usb_CDC_UNION                       = 0x06
-	usb_CDC_CS_INTERFACE                = 0x24
-	usb_CDC_CS_ENDPOINT                 = 0x25
-	usb_CDC_DATA_INTERFACE_CLASS        = 0x0A
-
-	usb_CDC_LINESTATE_DTR = 0x01
-	usb_CDC_LINESTATE_RTS = 0x02
+	endPoints = []uint32{
+		usb_CONTROL_ENDPOINT: usb_ENDPOINT_TYPE_CONTROL,
+		usb_CDC_ENDPOINT_ACM: (usb_ENDPOINT_TYPE_INTERRUPT | usbEndpointIn),
+		usb_CDC_ENDPOINT_OUT: (usb_ENDPOINT_TYPE_BULK | usbEndpointOut),
+		usb_CDC_ENDPOINT_IN:  (usb_ENDPOINT_TYPE_BULK | usbEndpointIn),
+		usb_HID_ENDPOINT_IN:  (usb_ENDPOINT_TYPE_DISABLE), // Interrupt In
+	}
 )
 
 // usbDeviceDescBank is the USB device endpoint descriptor.
-// typedef struct {
-// 	__IO USB_DEVICE_ADDR_Type      ADDR;        /**< \brief Offset: 0x000 (R/W 32) DEVICE_DESC_BANK Endpoint Bank, Adress of Data Buffer */
-// 	__IO USB_DEVICE_PCKSIZE_Type   PCKSIZE;     /**< \brief Offset: 0x004 (R/W 32) DEVICE_DESC_BANK Endpoint Bank, Packet Size */
-// 	__IO USB_DEVICE_EXTREG_Type    EXTREG;      /**< \brief Offset: 0x008 (R/W 16) DEVICE_DESC_BANK Endpoint Bank, Extended */
-// 	__IO USB_DEVICE_STATUS_BK_Type STATUS_BK;   /**< \brief Offset: 0x00A (R/W  8) DEVICE_DESC_BANK Enpoint Bank, Status of Bank */
-// 		 RoReg8                    Reserved1[0x5];
-//   } UsbDeviceDescBank;
 type usbDeviceDescBank struct {
 	ADDR      volatile.Register32
 	PCKSIZE   volatile.Register32
@@ -171,145 +189,68 @@ type usbDeviceDescriptor struct {
 	DeviceDescBank [2]usbDeviceDescBank
 }
 
-// typedef struct {
-// 	union {
-// 		uint8_t bmRequestType;
-// 		struct {
-// 			uint8_t direction : 5;
-// 			uint8_t type : 2;
-// 			uint8_t transferDirection : 1;
-// 		};
-// 	};
-// 	uint8_t bRequest;
-// 	uint8_t wValueL;
-// 	uint8_t wValueH;
-// 	uint16_t wIndex;
-// 	uint16_t wLength;
-// } USBSetup;
-type usbSetup struct {
-	bmRequestType uint8
-	bRequest      uint8
-	wValueL       uint8
-	wValueH       uint8
-	wIndex        uint16
-	wLength       uint16
+type USBSetup struct {
+	BmRequestType uint8
+	BRequest      uint8
+	WValueL       uint8
+	WValueH       uint8
+	WIndex        uint16
+	WLength       uint16
 }
 
-func newUSBSetup(data []byte) usbSetup {
-	u := usbSetup{}
-	u.bmRequestType = uint8(data[0])
-	u.bRequest = uint8(data[1])
-	u.wValueL = uint8(data[2])
-	u.wValueH = uint8(data[3])
-	u.wIndex = uint16(data[4]) | (uint16(data[5]) << 8)
-	u.wLength = uint16(data[6]) | (uint16(data[7]) << 8)
+func newUSBSetup(data []byte) USBSetup {
+	u := USBSetup{}
+	u.BmRequestType = uint8(data[0])
+	u.BRequest = uint8(data[1])
+	u.WValueL = uint8(data[2])
+	u.WValueH = uint8(data[3])
+	u.WIndex = uint16(data[4]) | (uint16(data[5]) << 8)
+	u.WLength = uint16(data[6]) | (uint16(data[7]) << 8)
 	return u
-}
-
-// USBCDC is the serial interface that works over the USB port.
-// To implement the USBCDC interface for a board, you must declare a concrete type as follows:
-//
-// 		type USBCDC struct {
-// 			Buffer *RingBuffer
-// 		}
-//
-// You can also add additional members to this struct depending on your implementation,
-// but the *RingBuffer is required.
-// When you are declaring the USBCDC for your board, make sure that you also declare the
-// RingBuffer using the NewRingBuffer() function:
-//
-//		USBCDC{Buffer: NewRingBuffer()}
-//
-
-// Read from the RX buffer.
-func (usbcdc *USBCDC) Read(data []byte) (n int, err error) {
-	// check if RX buffer is empty
-	size := usbcdc.Buffered()
-	if size == 0 {
-		return 0, nil
-	}
-
-	// Make sure we do not read more from buffer than the data slice can hold.
-	if len(data) < size {
-		size = len(data)
-	}
-
-	// only read number of bytes used from buffer
-	for i := 0; i < size; i++ {
-		v, _ := usbcdc.ReadByte()
-		data[i] = v
-	}
-
-	return size, nil
-}
-
-// Write data to the USBCDC.
-func (usbcdc *USBCDC) Write(data []byte) (n int, err error) {
-	for _, v := range data {
-		usbcdc.WriteByte(v)
-	}
-	return len(data), nil
-}
-
-// ReadByte reads a single byte from the RX buffer.
-// If there is no data in the buffer, returns an error.
-func (usbcdc *USBCDC) ReadByte() (byte, error) {
-	// check if RX buffer is empty
-	buf, ok := usbcdc.Buffer.Get()
-	if !ok {
-		return 0, errUSBCDCBufferEmpty
-	}
-	return buf, nil
-}
-
-// Buffered returns the number of bytes currently stored in the RX buffer.
-func (usbcdc *USBCDC) Buffered() int {
-	return int(usbcdc.Buffer.Used())
-}
-
-// Receive handles adding data to the UART's data buffer.
-// Usually called by the IRQ handler for a machine.
-func (usbcdc *USBCDC) Receive(data byte) {
-	usbcdc.Buffer.Put(data)
 }
 
 // sendDescriptor creates and sends the various USB descriptor types that
 // can be requested by the host.
-func sendDescriptor(setup usbSetup) {
-	switch setup.wValueH {
+func sendDescriptor(setup USBSetup) {
+	switch setup.WValueH {
 	case usb_CONFIGURATION_DESCRIPTOR_TYPE:
-		sendUSBPacket(0, usbDescriptor.Configuration, setup.wLength)
+		sendUSBPacket(0, usbDescriptor.Configuration, setup.WLength)
 		return
 	case usb_DEVICE_DESCRIPTOR_TYPE:
 		// composite descriptor
+		if (usbDescriptorConfig & usbDescriptorConfigHID) > 0 {
+			usbDescriptor = descriptorCDCHID
+		} else {
+			usbDescriptor = descriptorCDC
+		}
 		usbDescriptor.Configure(usb_VID, usb_PID)
-		sendUSBPacket(0, usbDescriptor.Device, setup.wLength)
+		sendUSBPacket(0, usbDescriptor.Device, setup.WLength)
 		return
 
 	case usb_STRING_DESCRIPTOR_TYPE:
-		switch setup.wValueL {
+		switch setup.WValueL {
 		case 0:
 			b := []byte{0x04, 0x03, 0x09, 0x04}
-			sendUSBPacket(0, b, setup.wLength)
+			sendUSBPacket(0, b, setup.WLength)
 
 		case usb_IPRODUCT:
 			b := make([]byte, (len(usb_STRING_PRODUCT)<<1)+2)
 			strToUTF16LEDescriptor(usb_STRING_PRODUCT, b)
-			sendUSBPacket(0, b, setup.wLength)
+			sendUSBPacket(0, b, setup.WLength)
 
 		case usb_IMANUFACTURER:
 			b := make([]byte, (len(usb_STRING_MANUFACTURER)<<1)+2)
 			strToUTF16LEDescriptor(usb_STRING_MANUFACTURER, b)
-			sendUSBPacket(0, b, setup.wLength)
+			sendUSBPacket(0, b, setup.WLength)
 
 		case usb_ISERIAL:
 			// TODO: allow returning a product serial number
-			sendZlp()
+			SendZlp()
 		}
 		return
 	case usb_HID_REPORT_TYPE:
-		if h, ok := usbDescriptor.HID[setup.wIndex]; ok {
-			sendUSBPacket(0, h, setup.wLength)
+		if h, ok := usbDescriptor.HID[setup.WIndex]; ok {
+			sendUSBPacket(0, h, setup.WLength)
 			return
 		}
 	case usb_DEVICE_QUALIFIER:
@@ -318,21 +259,103 @@ func sendDescriptor(setup usbSetup) {
 	}
 
 	// do not know how to handle this message, so return zero
-	sendZlp()
+	SendZlp()
 	return
 }
 
-// EnableHID enables HID. This function must be executed from the init().
-func EnableHID(callback func()) {
-	usbDescriptor = descriptorCDCHID
-	endPoints = []uint32{usb_ENDPOINT_TYPE_CONTROL,
-		(usb_ENDPOINT_TYPE_INTERRUPT | usbEndpointIn),
-		(usb_ENDPOINT_TYPE_BULK | usbEndpointOut),
-		(usb_ENDPOINT_TYPE_BULK | usbEndpointIn),
-		(usb_ENDPOINT_TYPE_INTERRUPT | usbEndpointIn)}
+func handleStandardSetup(setup USBSetup) bool {
+	switch setup.BRequest {
+	case usb_GET_STATUS:
+		buf := []byte{0, 0}
 
-	hidCallback = callback
+		if setup.BmRequestType != 0 { // endpoint
+			// TODO: actually check if the endpoint in question is currently halted
+			if isEndpointHalt {
+				buf[0] = 1
+			}
+		}
+
+		sendUSBPacket(0, buf, setup.WLength)
+		return true
+
+	case usb_CLEAR_FEATURE:
+		if setup.WValueL == 1 { // DEVICEREMOTEWAKEUP
+			isRemoteWakeUpEnabled = false
+		} else if setup.WValueL == 0 { // ENDPOINTHALT
+			isEndpointHalt = false
+		}
+		SendZlp()
+		return true
+
+	case usb_SET_FEATURE:
+		if setup.WValueL == 1 { // DEVICEREMOTEWAKEUP
+			isRemoteWakeUpEnabled = true
+		} else if setup.WValueL == 0 { // ENDPOINTHALT
+			isEndpointHalt = true
+		}
+		SendZlp()
+		return true
+
+	case usb_SET_ADDRESS:
+		return handleUSBSetAddress(setup)
+
+	case usb_GET_DESCRIPTOR:
+		sendDescriptor(setup)
+		return true
+
+	case usb_SET_DESCRIPTOR:
+		return false
+
+	case usb_GET_CONFIGURATION:
+		buff := []byte{usbConfiguration}
+		sendUSBPacket(0, buff, setup.WLength)
+		return true
+
+	case usb_SET_CONFIGURATION:
+		if setup.BmRequestType&usb_REQUEST_RECIPIENT == usb_REQUEST_DEVICE {
+			for i := 1; i < len(endPoints); i++ {
+				initEndpoint(uint32(i), endPoints[i])
+			}
+
+			usbConfiguration = setup.WValueL
+
+			SendZlp()
+			return true
+		} else {
+			return false
+		}
+
+	case usb_GET_INTERFACE:
+		buff := []byte{usbSetInterface}
+		sendUSBPacket(0, buff, setup.WLength)
+		return true
+
+	case usb_SET_INTERFACE:
+		usbSetInterface = setup.WValueL
+
+		SendZlp()
+		return true
+
+	default:
+		return true
+	}
 }
 
-// hidCallback is a variable that holds the callback when using HID.
-var hidCallback func()
+func EnableCDC(callback func(), callbackRx func([]byte), callbackSetup func(USBSetup) bool) {
+	usbDescriptorConfig |= usbDescriptorConfigCDC
+	endPoints[usb_CDC_ENDPOINT_ACM] = (usb_ENDPOINT_TYPE_INTERRUPT | usbEndpointIn)
+	endPoints[usb_CDC_ENDPOINT_OUT] = (usb_ENDPOINT_TYPE_BULK | usbEndpointOut)
+	endPoints[usb_CDC_ENDPOINT_IN] = (usb_ENDPOINT_TYPE_BULK | usbEndpointIn)
+	callbackUSBRx[usb_CDC_ENDPOINT_OUT] = callbackRx
+	callbackUSBTx[usb_CDC_ENDPOINT_IN] = callback
+	callbackUSBSetup[usb_CDC_ACM_INTERFACE] = callbackSetup // 0x02 (Communications and CDC Control)
+	callbackUSBSetup[usb_CDC_DATA_INTERFACE] = nil          // 0x0A (CDC-Data)
+}
+
+// EnableHID enables HID. This function must be executed from the init().
+func EnableHID(callback func(), callbackRx func([]byte), callbackSetup func(USBSetup) bool) {
+	usbDescriptorConfig |= usbDescriptorConfigHID
+	endPoints[usb_HID_ENDPOINT_IN] = (usb_ENDPOINT_TYPE_INTERRUPT | usbEndpointIn)
+	callbackUSBTx[usb_HID_ENDPOINT_IN] = callback
+	callbackUSBSetup[usb_HID_INTERFACE] = callbackSetup // 0x03 (HID - Human Interface Device)
+}
