+++ src/runtime/runtime_atmega.go
@@ -9,6 +9,7 @@ import (
 )
 
 func initUART() {
+	machine.InitSerial()
 	machine.Serial.Configure(machine.UARTConfig{})
 }
