+++ src/runtime/runtime_esp8266.go
@@ -48,6 +48,7 @@ func main() {
 	rom_i2c_writeReg(103, 4, 2, 145)
 
 	// Initialize UART.
+	machine.InitSerial()
 	machine.Serial.Configure(machine.UARTConfig{})
 
 	// Initialize timer. Bits:
