+++ src/runtime/runtime_stm32f103.go
@@ -11,6 +11,7 @@ import (
 func init() {
 	initCLK()
 
+	machine.InitSerial()
 	machine.Serial.Configure(machine.UARTConfig{})
 
 	initTickTimer(&machine.TIM4)
