+++ src/runtime/runtime_stm32f4.go
@@ -12,6 +12,7 @@ import (
 func init() {
 	initCLK()
 
+	machine.InitSerial()
 	machine.Serial.Configure(machine.UARTConfig{})
 
 	initTickTimer(&machine.TIM2)
