+++ src/runtime/runtime_stm32f405.go
@@ -156,6 +156,7 @@ func initCLK() {
 
 func initCOM() {
 	if machine.NUM_UART_INTERFACES > 0 {
+		machine.InitSerial()
 		machine.Serial.Configure(machine.UARTConfig{})
 	}
 }
