+++ src/runtime/runtime_stm32l0x1.go
@@ -15,6 +15,7 @@ const (
 func init() {
 	initCLK()
 
+	machine.InitSerial()
 	machine.Serial.Configure(machine.UARTConfig{})
 
 	initTickTimer(&machine.TIM21)
