+++ src/runtime/runtime_stm32l5x2.go
@@ -30,6 +30,7 @@ const (
 func init() {
 	initCLK()
 
+	machine.InitSerial()
 	machine.Serial.Configure(machine.UARTConfig{})
 
 	initTickTimer(&machine.TIM16)
